Nginx server add

+++ Docker/entrypoint.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-composer install --no-progress --no-interaction
+if [ ! -f "vendor/autoload.php" ]; then
+    composer install --no-progress --no-interaction
+fi
 
 if [ ! -f ".env" ]; then
     echo "Creating env file for env $APP_ENV"
